fixes: refactored group_controller and added yarn.lock
