+++ app/views/groups/_add_member_modal.html.erb
@@ -1,6 +1,4 @@
-<%= javascript_pack_tag 'application' %>
-
-<div id="addmemberModal" class="modal fade" role="dialog" data-controller="addMember" >
+<div id="addmemberModal" class="modal fade" role="dialog">
   <div class="modal-dialog primary-modal-dialog">
     <div class="modal-content">
       <div class="modal-header primary-modal-header">
